Changed newletter/page to ts and email collection to supa base
