+++ package.json
@@ -25,7 +25,10 @@
     "tsparticles-slim": "^2.12.0"
   },
   "devDependencies": {
+    "@types/node": "^20.19.43",
+    "@types/react": "^19.2.17",
     "eslint": "9.28.0",
-    "eslint-config-next": "15.3.3"
+    "eslint-config-next": "15.3.3",
+    "typescript": "^6.0.3"
   }
 }
